app-platform/app-mlflow/requirements.txt
mlflow==2.14.2
python-dotenv==1.0.1
streamlit==1.38.0
boto3